requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
google-genai==0.3.0
sentence-transformers==2.3.1
faiss-cpu>=1.7.4
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.0
tqdm==4.66.1
numpy==1.26.4
redis==5.0.1
psutil==5.9.8
pydantic==2.6.1
pydantic-settings==2.1.0
